первый коммит

+++ src/main/java/ru/yandex/practicum/filmorate/controller/Controller.java
@@ -8,22 +8,22 @@ import java.util.List;
 import java.util.Map;
 
 public abstract class Controller<T extends Entity> {
-    protected final Map<Integer, T> storage = new HashMap<>();
-    private int idCounter = 1;
+    protected final Map<Long, T> storage = new HashMap<>();
+    private Long idCounter = 1L;
 
     public List<T> getAll() {
         return new ArrayList<>(storage.values());
     }
 
     public T create(T entity) {
-        final int id = idCounter++;
+        final Long id = idCounter++;
         entity.setId(id);
         storage.put(id, entity);
         return storage.get(id);
     }
 
     public T update(T entity) {
-        final int id = entity.getId();
+        final Long id = entity.getId();
         storage.put(id, entity);
         return storage.get(id);
     }
